Fix: #8 refactor test directories
